Rename: 패키지 구조 변경 및 BaseEntity 테스트용 Setter 추가
