+++ src/main/java/ojosama/talkak/common/entity/BaseEntity.java
@@ -8,9 +8,13 @@ import lombok.Getter;
 
 @Getter
 @MappedSuperclass
-public class BaseEntity {
+public abstract class BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    public void createIdForTest(Long id) {
+        this.id = id;
+    }
 }
