+++ src/main/java/ojosama/talkak/recommendation/domain/Scores.java
@@ -1,0 +1,12 @@
+package ojosama.talkak.recommendation.domain;
+
+public record Scores(
+    Long categoryId,
+    Long videoId,
+    Float scores
+) {
+
+    public static Scores of(Long categoryId, Long videoId, Float scores) {
+        return new Scores(categoryId, videoId, scores);
+    }
+}
