+++ src/main/java/ojosama/talkak/recommendation/key/DefaultScoresKey.java
@@ -1,0 +1,21 @@
+package ojosama.talkak.recommendation.key;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum DefaultScoresKey {
+    SCORES("scores-default:%s");
+
+    private final String key;
+
+    public String generateKey(Long categoryId) {
+        return String.format(key, categoryId);
+    }
+
+    public Long getCategoryId(Long categoryId) {
+        return Long.parseLong(key.split(":")[1]);
+    }
+
+}
