+++ src/main/java/ojosama/talkak/recommendation/key/EventQueueKey.java
@@ -1,0 +1,12 @@
+package ojosama.talkak.recommendation.key;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum EventQueueKey {
+    QUEUE("event_queue");
+
+    private final String key;
+}
