--- src/main/java/ojosama/talkak/redis/domain/Scores.java
@@ -1,12 +1,0 @@
-package ojosama.talkak.redis.domain;
-
-public record Scores(
-    Long categoryId,
-    Long videoId,
-    Float scores
-) {
-
-    public static Scores of(Long categoryId, Long videoId, Float scores) {
-        return new Scores(categoryId, videoId, scores);
-    }
-}
