--- src/main/java/ojosama/talkak/redis/innerkey/EventQueueHashKey.java
@@ -1,14 +1,0 @@
-package ojosama.talkak.redis.innerkey;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public enum EventQueueHashKey {
-    MEMBER_ID("member_id"),
-    CATEGORY_ID("category_id"),
-    LAST_UPDATED_AT("last_updated_at");
-
-    private final String key;
-}
