--- src/main/java/ojosama/talkak/redis/innerkey/ReactionHashKey.java
@@ -1,13 +1,0 @@
-package ojosama.talkak.redis.innerkey;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public enum ReactionHashKey {
-    WATCHED("watched"),
-    LIKED("liked");
-
-    private final String key;
-}
