--- src/main/java/ojosama/talkak/redis/innerkey/UserMetaHashKey.java
@@ -1,12 +1,0 @@
-package ojosama.talkak.redis.innerkey;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public enum UserMetaHashKey {
-    LAST_UPDATED_AT("last_updated_at");
-
-    private final String key;
-}
