--- src/main/java/ojosama/talkak/redis/innerkey/VideoHashKey.java
@@ -1,14 +1,0 @@
-package ojosama.talkak.redis.innerkey;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public enum VideoHashKey {
-    CREATED_AT("created_at"),
-    VIEW_COUNT("view_count"),
-    LIKE_COUNT("like_count");
-
-    private final String key;
-}
