--- src/main/java/ojosama/talkak/redis/key/EventQueueKey.java
@@ -1,12 +1,0 @@
-package ojosama.talkak.redis.key;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public enum EventQueueKey {
-    QUEUE("event_queue");
-
-    private final String key;
-}
