+++ src/main/java/ojosama/talkak/video/service/VideoService.java
@@ -7,8 +7,8 @@ import ojosama.talkak.common.exception.code.MemberError;
 import ojosama.talkak.common.exception.code.VideoError;
 import ojosama.talkak.member.domain.Member;
 import ojosama.talkak.member.repository.MemberRepository;
-import ojosama.talkak.redis.domain.VideoInfo;
-import ojosama.talkak.redis.repository.VideoInfoRepository;
+import ojosama.talkak.recommendation.domain.VideoInfo;
+import ojosama.talkak.recommendation.repository.VideoInfoRepository;
 import ojosama.talkak.reaction.service.ReactionService;
 import ojosama.talkak.video.domain.Video;
 import ojosama.talkak.video.repository.VideoRepository;
