+++ src/test/java/ojosama/talkak/redis/repository/VideoInfoConcurrencyTest.java
@@ -6,9 +6,10 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import lombok.extern.slf4j.Slf4j;
+import ojosama.talkak.recommendation.repository.VideoInfoRepository;
 import ojosama.talkak.redis.config.RecommendTestContainerConfig;
 import ojosama.talkak.redis.config.RedisTestContainerConfig;
-import ojosama.talkak.redis.domain.VideoInfo;
+import ojosama.talkak.recommendation.domain.VideoInfo;
 import ojosama.talkak.reaction.service.ReactionService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
